Revert "layout/social-page start-up setting "
