--- src/pages/test/pet_profile.tsx
@@ -1,7 +1,0 @@
-export default function Social() {
-  return (
-    <main className="flex flex-col items-center justify-between p-24 min-h-screen bg-slate-400">
-      我是寵物檔案頁面
-    </main>
-  );
-}
