--- src/pages/test/social.tsx
@@ -1,12 +1,0 @@
-import Navbar from "@/components/Navbar";
-
-export default function Social() {
-  return (
-    <>
-      <Navbar></Navbar>
-      <main className="flex flex-col items-center justify-between p-24 min-h-screen bg-slate-400">
-        我是社群頁面
-      </main>
-    </>
-  );
-}
