--- src/pages/test/upload.tsx
@@ -1,9 +1,0 @@
-import Upload from "@/components/test/Upload";
-
-export default function Social() {
-  return (
-    <main className="flex flex-col items-center justify-between p-24 min-h-screen bg-slate-400">
-      <Upload></Upload>
-    </main>
-  );
-}
