+++ src/styles/globals.css
@@ -1,12 +1,3 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-
-@layer components {
-  .inner {
-    @apply mx-auto max-w-[1344px];
-  }
-  .outter {
-    @apply mx-auto max-w-[1440px];
-  }
-}
